requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
nltk>=3.8
requests>=2.31.0
pyyaml>=6.0
pathlib2>=2.3.7
numpy>=1.24.0